Refactor vertices.
