+++ src/stk/_internal/topology_graphs/polymer/__init__.py
@@ -6,3 +6,5 @@ Polymer and Small Molecules
 #. :class:`.NCore`
 
 """
+
+from . import vertices  # noqa
